requirements.txt
Django~=3.1.7
pytest
Pillow~=8.2.0
djangorestframework~=3.12.4